+++ src/cli/mod.rs
@@ -90,9 +90,6 @@ impl Cli {
         let rpc_endpoint = client_config.rpc.endpoint.to_string();
         let store = SqliteStore::new((&client_config).into()).map_err(ClientError::StoreError)?;
         let rng = get_random_coin();
-        let _executor_store =
-            miden_client::store::sqlite_store::SqliteStore::new((&client_config).into())
-                .map_err(ClientError::StoreError)?;
 
         let client: Client<TonicRpcClient, RpoRandomCoin, SqliteStore> =
             Client::new(TonicRpcClient::new(&rpc_endpoint), rng, store, in_debug_mode);