+++ src/store/sqlite_store/mod.rs
@@ -4,7 +4,7 @@ use core::cell::{RefCell, RefMut};
 use miden_objects::{
     accounts::{Account, AccountId, AccountStub},
     crypto::merkle::{InOrderIndex, MmrPeaks},
-    notes::{NoteId, NoteTag},
+    notes::NoteTag,
     transaction::TransactionId,
     BlockHeader, Digest, Word,
 };
@@ -175,10 +175,6 @@ impl Store for SqliteStore {
         note_filter: NoteFilter,
     ) -> Result<Vec<OutputNoteRecord>, StoreError> {
         self.get_output_notes(note_filter)
-    }
-
-    fn get_input_note(&self, note_id: NoteId) -> Result<InputNoteRecord, StoreError> {
-        self.get_input_note(note_id)
     }
 
     fn insert_input_note(&self, note: &InputNoteRecord) -> Result<(), StoreError> {